+++ app/Ajax/App/Tontine/Charge.php
@@ -3,7 +3,7 @@
 namespace App\Ajax\App\Tontine;
 
 use Siak\Tontine\Service\LocaleService;
-use Siak\Tontine\Service\Planning\ChargeService;
+use Siak\Tontine\Service\Tontine\ChargeService;
 use Siak\Tontine\Validation\Tontine\ChargeValidator;
 use App\Ajax\App\Faker;
 use App\Ajax\CallableClass;
