+++ app/Providers/SiakServiceProvider.php
@@ -23,12 +23,12 @@ use Siak\Tontine\Service\Meeting\Pool\PoolService as MeetingPoolService;
 use Siak\Tontine\Service\Meeting\Pool\RemitmentService;
 use Siak\Tontine\Service\Meeting\SessionService as MeetingSessionService;
 use Siak\Tontine\Service\Meeting\SummaryService as MeetingSummaryService;
-use Siak\Tontine\Service\Planning\ChargeService;
+use Siak\Tontine\Service\Planning\PoolService as TontinePoolService;
 use Siak\Tontine\Service\Planning\RoundService;
 use Siak\Tontine\Service\Planning\SessionService as PlanningSessionService;
 use Siak\Tontine\Service\Planning\SubscriptionService;
 use Siak\Tontine\Service\Planning\SummaryService as PlanningSummaryService;
-use Siak\Tontine\Service\Tontine\PoolService as TontinePoolService;
+use Siak\Tontine\Service\Tontine\ChargeService;
 use Siak\Tontine\Service\Tontine\MemberService;
 use Siak\Tontine\Service\Tontine\TontineService;
 use Siak\Tontine\Validation\Meeting\DebtValidator;
