+++ src/Service/Balance/MemberService.php
@@ -14,7 +14,7 @@ use Siak\Tontine\Model\Session;
 use Siak\Tontine\Model\Subscription;
 use Siak\Tontine\Service\LocaleService;
 use Siak\Tontine\Service\TenantService;
-use Siak\Tontine\Service\Tontine\PoolService;
+use Siak\Tontine\Service\Planning\PoolService;
 
 class MemberService
 {
